Refactor note preview component
