+++ app/views/note/preview/NotePreviewStyles.js
@@ -1,0 +1,36 @@
+export default Styles = {
+  container: {
+    flex: 1,
+    justifyContent: 'center',
+    alignItems: 'center',
+    backgroundColor: '#F5FCFF',
+  },
+  markdownStyles: {
+    container: {
+      paddingLeft: 10
+    },
+    heading1: {
+      fontSize: 24,
+      fontWeight: '600',
+      color: '#222222',
+    },
+    link: {
+      color: 'red',
+    },
+    mail_to: {
+      color: 'orange',
+    },
+    text: {
+      color: '#555555',
+    },
+    code: {
+      backgroundColor: '#f0f0f0',
+      marginTop: 5,
+      marginBottom: 5
+    },
+    blockquote: {
+      backgroundColor: '#f8f8f8',
+      padding: 5
+    }
+  },
+}
